refactor: centralize state management with Zustand
