+++ src/stores/timers.ts
@@ -1,29 +1,10 @@
-import React, { useEffect, useState } from "react";
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
-import {
-  loadOfflineData,
-  updateOfflineData,
-  syncWithServer,
-} from "@/utils/offline";
+import { Timer } from "@/types";
 
 const generateId = () =>
   (crypto as { randomUUID?: () => string }).randomUUID?.() ||
   `${Date.now()}-${Math.random().toString(36).slice(2)}`;
-
-export interface Timer {
-  id: string;
-  title: string;
-  color: number;
-  baseDuration: number;
-  duration: number;
-  remaining: number;
-  isRunning: boolean;
-  isPaused: boolean;
-  startTime?: number;
-  lastTick?: number;
-  pauseStart?: number;
-}
 
 interface TimersState {
   timers: Timer[];
@@ -202,59 +183,3 @@ export const useTimers = create<TimersState>()(
     { name: "timers" },
   ),
 );
-
-export const TimersProvider: React.FC<{ children: React.ReactNode }> = ({
-  children,
-}) => {
-  const setTimers = useTimers((s) => s.setTimers);
-  const timers = useTimers((s) => s.timers);
-  const [loaded, setLoaded] = useState(false);
-
-  useEffect(() => {
-    const load = async () => {
-      const offline = loadOfflineData();
-      if (offline) setTimers(offline.timers || []);
-      const synced = await syncWithServer();
-      setTimers(synced.timers || []);
-      setLoaded(true);
-    };
-    load();
-  }, [setTimers]);
-
-  useEffect(() => {
-    if (!loaded) return;
-
-    // Only sync for non-tick updates (debounced approach)
-    const hasRunningTimers = timers.some((t) => t.isRunning && !t.isPaused);
-
-    const save = async () => {
-      try {
-        updateOfflineData({ timers });
-        if (navigator.onLine) {
-          await fetch("/api/timers", {
-            method: "PUT",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify(timers),
-          });
-
-          // Only sync with server for significant changes, not tick updates
-          if (!hasRunningTimers) {
-            await syncWithServer();
-          }
-        }
-      } catch (err) {
-        console.error("Error saving timers", err);
-      }
-    };
-
-    // Debounce: Don't sync immediately if timers are running (frequent ticks)
-    if (hasRunningTimers) {
-      const timeoutId = setTimeout(save, 5000); // Sync every 5 seconds max when running
-      return () => clearTimeout(timeoutId);
-    } else {
-      save(); // Immediate sync for non-running states (start/stop/pause actions)
-    }
-  }, [timers, loaded]);
-
-  return <>{children}</>;
-};
